FIR/UAST: commonize declaration conversion
