+++ uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/BaseKotlinConverter.kt
@@ -6,15 +6,12 @@
 package org.jetbrains.uast.kotlin
 
 import com.intellij.openapi.components.ServiceManager
-import com.intellij.psi.PsiComment
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiEnumConstant
-import com.intellij.psi.PsiParameterList
+import com.intellij.psi.*
 import com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.kotlin.asJava.LightClassUtil
 import org.jetbrains.kotlin.asJava.classes.KtLightClass
-import org.jetbrains.kotlin.asJava.elements.KtLightElementBase
-import org.jetbrains.kotlin.asJava.elements.KtLightField
+import org.jetbrains.kotlin.asJava.elements.*
+import org.jetbrains.kotlin.asJava.findFacadeClass
 import org.jetbrains.kotlin.asJava.toLightClass
 import org.jetbrains.kotlin.kdoc.psi.impl.KDocLink
 import org.jetbrains.kotlin.kdoc.psi.impl.KDocName
@@ -23,12 +20,13 @@ import org.jetbrains.kotlin.kdoc.psi.impl.KDocTag
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
+import org.jetbrains.kotlin.utils.addToStdlib.safeAs
 import org.jetbrains.uast.*
-import org.jetbrains.uast.kotlin.psi.UastFakeLightPrimaryConstructor
-import org.jetbrains.uast.kotlin.psi.UastKotlinPsiParameter
-import org.jetbrains.uast.kotlin.psi.UastKotlinPsiVariable
+import org.jetbrains.uast.kotlin.psi.*
 
 interface BaseKotlinConverter {
+    val languagePlugin: UastLanguagePlugin
+
     fun unwrapElements(element: PsiElement?): PsiElement? = when (element) {
         is KtValueArgumentList -> unwrapElements(element.parent)
         is KtValueArgument -> unwrapElements(element.parent)
@@ -58,7 +56,161 @@ interface BaseKotlinConverter {
         element: PsiElement,
         givenParent: UElement?,
         requiredTypes: Array<out Class<out UElement>>
-    ): UElement?
+    ): UElement? {
+        val original = element.originalElement
+
+        fun <P : PsiElement> build(ctor: (P, UElement?) -> UElement): () -> UElement? = {
+            @Suppress("UNCHECKED_CAST")
+            ctor(original as P, givenParent)
+        }
+
+        fun <P : PsiElement, K : KtElement> buildKt(ktElement: K, ctor: (P, K, UElement?) -> UElement): () -> UElement? = {
+            @Suppress("UNCHECKED_CAST")
+            ctor(original as P, ktElement, givenParent)
+        }
+
+        fun <P : PsiElement, K : KtElement> buildKtOpt(ktElement: K?, ctor: (P, K?, UElement?) -> UElement): () -> UElement? = {
+            @Suppress("UNCHECKED_CAST")
+            ctor(original as P, ktElement, givenParent)
+        }
+
+        fun Array<out Class<out UElement>>.convertToUField(original: PsiField, kotlinOrigin: KtElement?): UElement? =
+            if (original is PsiEnumConstant)
+                el<UEnumConstant>(buildKtOpt(kotlinOrigin, ::KotlinUEnumConstant))
+            else
+                el<UField>(buildKtOpt(kotlinOrigin, ::KotlinUField))
+
+        return with(requiredTypes) {
+            when (original) {
+                is KtLightMethod -> {
+                    el<UMethod>(build(KotlinUMethod::create))
+                }
+                is UastFakeLightMethod -> {
+                    el<UMethod> {
+                        val ktFunction = original.original
+                        if (ktFunction.isLocal)
+                            convertDeclaration(ktFunction, givenParent, requiredTypes)
+                        else
+                            KotlinUMethodWithFakeLightDelegate(ktFunction, original, givenParent)
+                    }
+                }
+                is UastFakeLightPrimaryConstructor -> {
+                    convertFakeLightConstructorAlternatives(original, givenParent, requiredTypes).firstOrNull()
+                }
+
+                is KtLightClass -> {
+                    when (original.kotlinOrigin) {
+                        is KtEnumEntry -> el<UEnumConstant> {
+                            convertEnumEntry(original.kotlinOrigin as KtEnumEntry, givenParent)
+                        }
+                        else -> el<UClass> { KotlinUClass.create(original, givenParent) }
+                    }
+                }
+
+                is KtLightField -> {
+                    convertToUField(original, original.kotlinOrigin)
+                }
+                is KtLightFieldForSourceDeclarationSupport -> {
+                    // KtLightFieldForDecompiledDeclaration is not a KtLightField
+                    convertToUField(original, original.kotlinOrigin)
+                }
+                is KtLightParameter -> {
+                    el<UParameter>(buildKtOpt(original.kotlinOrigin, ::KotlinUParameter))
+                }
+                is UastKotlinPsiParameter -> {
+                    el<UParameter>(buildKt(original.ktParameter, ::KotlinUParameter))
+                }
+                is UastKotlinPsiParameterBase<*> -> {
+                    el<UParameter> {
+                        original.ktOrigin.safeAs<KtTypeReference>()?.let { convertReceiverParameter(it) }
+                    }
+                }
+                is UastKotlinPsiVariable -> {
+                    el<ULocalVariable>(buildKt(original.ktElement, ::KotlinULocalVariable))
+                }
+
+                is KtEnumEntry -> {
+                    el<UEnumConstant> {
+                        convertEnumEntry(original, givenParent)
+                    }
+                }
+                is KtClassOrObject -> {
+                    convertClassOrObject(original, givenParent, this).firstOrNull()
+                }
+
+                is KtFunction -> {
+                    if (original.isLocal) {
+                        el<ULambdaExpression> {
+                            val parent = original.parent
+                            if (parent is KtLambdaExpression) {
+                                KotlinULambdaExpression(parent, givenParent) // your parent is the ULambdaExpression
+                            } else if (original.name.isNullOrEmpty()) {
+                                createLocalFunctionLambdaExpression(original, givenParent)
+                            } else {
+                                val uDeclarationsExpression = createLocalFunctionDeclaration(original, givenParent)
+                                val localFunctionVar = uDeclarationsExpression.declarations.single() as KotlinLocalFunctionUVariable
+                                localFunctionVar.uastInitializer
+                            }
+                        }
+                    } else {
+                        el<UMethod> {
+                            val lightMethod = LightClassUtil.getLightClassMethod(original)
+                            if (lightMethod != null)
+                                convertDeclaration(lightMethod, givenParent, requiredTypes)
+                            else {
+                                val ktLightClass = getLightClassForFakeMethod(original) ?: return null
+                                KotlinUMethodWithFakeLightDelegate(original, ktLightClass, givenParent)
+                            }
+                        }
+                    }
+                }
+
+                is KtPropertyAccessor -> {
+                    el<UMethod> {
+                        val lightMethod = LightClassUtil.getLightClassAccessorMethod(original) ?: return null
+                        convertDeclaration(lightMethod, givenParent, requiredTypes)
+                    }
+                }
+                is KtProperty -> {
+                    if (original.isLocal) {
+                        convertPsiElement(original, givenParent, requiredTypes)
+                    } else {
+                        convertNonLocalProperty(original, givenParent, requiredTypes).firstOrNull()
+                    }
+                }
+                is KtParameter -> {
+                    convertParameter(original, givenParent, this).firstOrNull()
+                }
+
+                is KtFile -> {
+                    convertKtFile(original, givenParent, this).firstOrNull()
+                }
+                is FakeFileForLightClass -> {
+                    el<UFile> { KotlinUFile(original.navigationElement, languagePlugin) }
+                }
+
+                is KtAnnotationEntry -> {
+                    el<UAnnotation>(build(::convertAnnotation))
+                }
+                is KtCallExpression -> {
+                    if (requiredTypes.isAssignableFrom(KotlinUNestedAnnotation::class.java) &&
+                        !requiredTypes.isAssignableFrom(UCallExpression::class.java)
+                    ) {
+                        el<UAnnotation> { KotlinUNestedAnnotation.create(original, givenParent) }
+                    } else null
+                }
+                is KtLightAnnotationForSourceEntry -> {
+                    convertDeclarationOrElement(original.kotlinOrigin, givenParent, requiredTypes)
+                }
+                is KtDelegatedSuperTypeEntry -> {
+                    el<KotlinSupertypeDelegationUExpression> {
+                        KotlinSupertypeDelegationUExpression(original, givenParent)
+                    }
+                }
+                else -> null
+            }
+        }
+    }
 
     fun convertDeclarationOrElement(
         element: PsiElement,
@@ -68,6 +220,19 @@ interface BaseKotlinConverter {
         return if (element is UElement) element
         else convertDeclaration(element, givenParent, expectedTypes)
             ?: convertPsiElement(element, givenParent, expectedTypes)
+    }
+
+    fun convertKtFile(
+        element: KtFile,
+        givenParent: UElement?,
+        requiredTypes: Array<out Class<out UElement>>
+    ): Sequence<UElement> {
+        return requiredTypes.accommodate(
+            // File
+            alternative { KotlinUFile(element, languagePlugin) },
+            // Facade
+            alternative { element.findFacadeClass()?.let { KotlinUClass.create(it, givenParent) } }
+        )
     }
 
     fun convertClassOrObject(
@@ -101,12 +266,12 @@ interface BaseKotlinConverter {
         )
     }
 
-    fun getLightClassForFakeMethod(original: KtFunction): KtLightClass? {
+    private fun getLightClassForFakeMethod(original: KtFunction): KtLightClass? {
         if (original.isLocal) return null
         return getContainingLightClass(original)
     }
 
-    fun convertToPropertyAlternatives(
+    private fun convertToPropertyAlternatives(
         methods: LightClassUtil.PropertyAccessorsPsiMethods?,
         givenParent: UElement?
     ): Array<UElementAlternative<*>> {
@@ -129,7 +294,32 @@ interface BaseKotlinConverter {
         )
     }
 
-    fun convertEnumEntry(original: KtEnumEntry, givenParent: UElement?): UElement? {
+    fun convertParameter(
+        element: KtParameter,
+        givenParent: UElement?,
+        requiredTypes: Array<out Class<out UElement>>
+    ): Sequence<UElement> =
+        requiredTypes.accommodate(
+            alternative uParam@{
+                val lightMethod = when (val ownerFunction = element.ownerFunction) {
+                    is KtFunction -> LightClassUtil.getLightClassMethod(ownerFunction)
+                        ?: getLightClassForFakeMethod(ownerFunction)
+                            ?.takeIf { !it.isAnnotationType }
+                            ?.let { UastFakeLightMethod(ownerFunction, it) }
+                    is KtPropertyAccessor -> LightClassUtil.getLightClassAccessorMethod(ownerFunction)
+                    else -> null
+                } ?: return@uParam null
+                val lightParameter = lightMethod.parameterList.parameters.find { it.name == element.name } ?: return@uParam null
+                KotlinUParameter(lightParameter, element, givenParent)
+            },
+            alternative catch@{
+                val uCatchClause = element.parent?.parent?.safeAs<KtCatchClause>()?.toUElementOfType<UCatchClause>() ?: return@catch null
+                uCatchClause.parameters.firstOrNull { it.sourcePsi == element }
+            },
+            *convertToPropertyAlternatives(LightClassUtil.getLightClassPropertyMethods(element), givenParent)
+        )
+
+    private fun convertEnumEntry(original: KtEnumEntry, givenParent: UElement?): UElement? {
         return LightClassUtil.getLightClassBackingField(original)?.let { psiField ->
             if (psiField is KtLightField && psiField is PsiEnumConstant) {
                 KotlinUEnumConstant(psiField, psiField.kotlinOrigin, givenParent)
@@ -139,7 +329,7 @@ interface BaseKotlinConverter {
         }
     }
 
-    fun convertReceiverParameter(receiver: KtTypeReference): UParameter? {
+    private fun convertReceiverParameter(receiver: KtTypeReference): UParameter? {
         val call = (receiver.parent as? KtCallableDeclaration) ?: return null
         if (call.receiverTypeReference != receiver) return null
         return call.toUElementOfType<UMethod>()?.uastParameters?.firstOrNull()
