--- uast/uast-kotlin-fir/testData/legacyIdentifiers/CommentOwners.identifiers.fir.txt
@@ -1,35 +1,0 @@
-topLevelProperty -> UField (name = topLevelProperty)
-Int -> USimpleNameReferenceExpression (identifier = Int)
-topLevelFun -> UMethod (name = topLevelFun)
-Unit -> USimpleNameReferenceExpression (identifier = Unit)
-TODO -> UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
-TopLevelClass -> UClass (name = TopLevelClass)
-constructor -> UMethod (name = TopLevelClass)
-t -> UParameter (name = t)
-Int -> USimpleNameReferenceExpression (identifier = Int)
-classLevelProperty -> UField (name = classLevelProperty)
-Int -> USimpleNameReferenceExpression (identifier = Int)
-classLevelMethod -> UMethod (name = classLevelMethod)
-Unit -> USimpleNameReferenceExpression (identifier = Unit)
-TODO -> UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
-NestedClass -> UClass (name = NestedClass)
-funPlainCall -> UMethod (name = funPlainCall)
-a -> UParameter (name = a)
-Int -> USimpleNameReferenceExpression (identifier = Int)
-Unit -> USimpleNameReferenceExpression (identifier = Unit)
-TODO -> UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
-funNamedArgumentsCall -> UMethod (name = funNamedArgumentsCall)
-a -> UParameter (name = a)
-Int -> USimpleNameReferenceExpression (identifier = Int)
-Unit -> USimpleNameReferenceExpression (identifier = Unit)
-TODO -> UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
-func -> UMethod (name = func)
-a -> UParameter (name = a)
-Int -> USimpleNameReferenceExpression (identifier = Int)
-    funPlainCall -> UCallExpression (kind = UastCallKind(name='method_call'), argCount = 1))
-    funNamedArgumentsCall -> UCallExpression (kind = UastCallKind(name='method_call'), argCount = 1))
-    a -> USimpleNameReferenceExpression (identifier = a)
-    localValueDefinition -> ULocalVariable (name = localValueDefinition)
-MyBooleanEnum -> UClass (name = MyBooleanEnum)
-TRUE -> UastEmptyExpression
-FALSE -> UastEmptyExpression
