+++ uast/uast-kotlin-fir/testData/legacyIdentifiers/CommentOwners.identifiers.txt
