+++ uast/uast-kotlin-fir/testData/legacyIdentifiers/EnumValuesConstructors.identifiers.txt
